+++ src/Data/Bifunctor/Biff.hs
@@ -1,4 +1,6 @@
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -7,7 +9,7 @@
 {-# LANGUAGE Safe #-}
 
 -- |
--- Copyright   :  (C) 2008-2016 Edward Kmett
+-- Copyright   :  (C) 2008-2021 Edward Kmett
 -- License     :  BSD-style (see the file LICENSE)
 -- Maintainer  :  Edward Kmett <ekmett@gmail.com>
 -- Stability   :  provisional
@@ -26,10 +28,12 @@ import GHC.Generics
 
 -- | Compose two 'Functor's on the inside of a 'Bifunctor'.
 newtype Biff p f g a b = Biff { runBiff :: p (f a) (g b) }
-  deriving ( Eq, Ord, Show, Read, Data
-           , Generic
-           )
-deriving instance Functor (p (f a)) => Generic1 (Biff p f g a)
+  deriving stock (Eq, Ord, Show, Read, Data, Generic)
+
+deriving stock instance Functor (p (f a)) => Generic1 (Biff p f g a)
+deriving stock instance (Functor (p (f a)), Functor g) => Functor (Biff p f g a)
+deriving stock instance (Foldable (p (f a)), Foldable g) => Foldable (Biff p f g a)
+deriving stock instance (Traversable (p (f a)), Traversable g) => Traversable (Biff p f g a)
 
 instance (Eq2 p, Eq1 f, Eq1 g, Eq a) => Eq1 (Biff p f g a) where
   liftEq = liftEq2 (==)
@@ -70,10 +74,6 @@ instance (Bifunctor p, Functor f, Functor g) => Bifunctor (Biff p f g) where
   bimap f g = Biff . bimap (fmap f) (fmap g) . runBiff
   {-# INLINE bimap #-}
 
-instance (Bifunctor p, Functor g) => Functor (Biff p f g a) where
-  fmap f = Biff . second (fmap f) . runBiff
-  {-# INLINE fmap #-}
-
 instance (Biapplicative p, Applicative f, Applicative g) => Biapplicative (Biff p f g) where
   bipure a b = Biff (bipure (pure a) (pure b))
   {-# INLINE bipure #-}
@@ -81,17 +81,9 @@ instance (Biapplicative p, Applicative f, Applicative g) => Biapplicative (Biff 
   Biff fg <<*>> Biff xy = Biff (bimap (<*>) (<*>) fg <<*>> xy)
   {-# INLINE (<<*>>) #-}
 
-instance (Bifoldable p, Foldable g) => Foldable (Biff p f g a) where
-  foldMap f = bifoldMap (const mempty) (foldMap f) . runBiff
-  {-# INLINE foldMap #-}
-
 instance (Bifoldable p, Foldable f, Foldable g) => Bifoldable (Biff p f g) where
   bifoldMap f g = bifoldMap (foldMap f) (foldMap g) . runBiff
   {-# INLINE bifoldMap #-}
-
-instance (Bitraversable p, Traversable g) => Traversable (Biff p f g a) where
-  traverse f = fmap Biff . bitraverse pure (traverse f) . runBiff
-  {-# INLINE traverse #-}
 
 instance (Bitraversable p, Traversable f, Traversable g) => Bitraversable (Biff p f g) where
   bitraverse f g = fmap Biff . bitraverse (traverse f) (traverse g) . runBiff