+++ src/Data/Bifunctor/Product.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE PolyKinds #-}
@@ -30,7 +31,7 @@ import Prelude hiding ((.),id)
 
 -- | Form the product of two bifunctors
 data Product f g a b = Pair (f a b) (g a b)
-  deriving ( Eq, Ord, Show, Read, Data, Generic, Generic1 )
+  deriving ( Eq, Ord, Show, Read, Data, Generic, Generic1, Functor, Foldable, Traversable )
 
 instance (Eq2 f, Eq2 g, Eq a) => Eq1 (Product f g a) where
   liftEq = liftEq2 (==)